refactor: render.ts를 목록, 모달, 페이지 렌더링 모듈로 분리
