+++ src/render/index.ts
@@ -1,0 +1,13 @@
+export {
+  createMovieItemTemplate,
+  createSkeletonItemTemplate,
+  removeSkeletonItem,
+  removeTopRatedMovieSkeleton,
+  renderListTitle,
+  renderMovies,
+  renderSkeletonItems,
+  renderTopRatedMovie,
+  updateEmptyListAlert,
+} from "./list";
+export { initModal, openMovieModal } from "./modal";
+export { renderMoviePage } from "./page";
